Installed react via npx, set up firebase, create Navbar and Home.js files, and import both to the App.js file
App.js is the root of the component and it is at the top while all the other components are imported to that root component. It is the first component that gets rendered to the DOM. The Navbar and Home components are exported to the root component. In the Navbar file a className named navbar and links are used for styling that block of the nav and the div. In the div block two links are created a Home link with just a / and a New Blog link with a /create link name