Click events with onClick
The file starts with creating the functions on top of the return, making it a const than giving it a name which is 'handleClick' then it is followed by an = whit a () followed by an arrow pointing at an object which is siginified by two brackets. The on call function is placed between two brackets which indicates its an object and that its from html to js zone the we move out into html area which is the button, so onClick which is an html button = the function js object function handleClick. The onCall function can pass string values by placing it in the () between '' and the const handleClick will = an anonymous function that passes a variable that contains the string value. To link const function to button you use onClick =ing the name of the function. Passing an argument to a function requires you to use an anonymous function. Can't pass the value directly cause that will invoke the function at run time not when it is invoked by a button. The outside {} indicates a dynamic value.