Outputting useState lists, using map
map is used to cycle through all array using js in the template and output the template for  each one. The map method in js which cycles the array and something to each item in the array. return a bit of template for each item as we iterate through each one and then that template is going to the output to the browser. In the template a the blogs property with a map method is written in a template.