props & reusablity
Reusing bits of templates and make it its own reusable component. You can drop the component into other components and don't have to repeat self. Pass in different data into the reusable component everytime we use it, data can be passed into these external data in the form of props. The data is different but the structure is the same. BlogList file created to house the reusable structure and data is passed through it via props. Add <BlogList/> on the Home file then import BlogList to that file. We are passing data from parent to child component. Make property name in <BlogList> as a prop.