handleDelete button to delete blogs, npm run build & deployed to firebase, functions as props
Create a button in the bloglist.js file, the button has an onClick which = a dynamic object using bracket and in the bracket there is an anonymous function pointed at handleDelete method with blog.id as its paramter. A 'delete blog' is written between the button element. In the Home.js file a const handleDelete function is created and within the BlogList tag 'handleDelete={handleDelete}' is inserted if you don't do that you will get a 'handleDelete is not a function' error. In the blogList file pass handleDelete as a props byt typing 'const handleDelete = props.handleDelete'