useEffect function, add json data to db.json file and then db.json file to data folder, add fetch inside useEffect to fetch data
useEffect runs a function every render of the component, the component renders initally when it first loads and it renders all of this to the dom but it also does it when there is a state change and when a value of a state is one of the useEffect dependency. In this code there is only an empty dependency cause there is an empty bracket and fetch is in the bracket of the useEffect, so it will run once or fetch the data then store it in the setBlogs const then in the template it will be expressed. The useEffect runs on every render.